backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0.post1
sqlmodel==0.0.16
pydantic-settings==2.2.1
python-multipart==0.0.6
pillow==10.2.0
pytesseract==0.3.10
